First crack at the GTK+ version of the Disconnected Client for Linux based operating systems. Will need to implement a shim similar to the MiniIms for this as there is no way to natively expose javascript objects
