+++ GtkClient/WebkitSettings.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace GtkClient
+{
+	public class WebkitSettings : WebKit.WebSettings
+	{
+		public WebkitSettings (String userAgent)
+		{
+			base.SetProperty ("user-agent", new GLib.Value (userAgent));
+
+		}
+	}
+}
+
